Fixing Parser Unit Tests (#372)
Parser unit tests would always pass, regardless of whether the answer
was correct or not, due to incorrect pattern matching.
- Now each unit tests expects a concrete value produced by a given
parser (previously was always a string)
- Bumped the precision of our float literal parser from "Float" type to
"Double" type, as it was not precise enough for some of the test cases.
- There exists BigDecimal type in Java which is able to represent
arbitrary-precision decimal numbers, however I chose not to for now in
line with not overthinking it too much!