Refactor UnregisteredOperation and current usages. (#380)
Slight rework of UnregisteredOperation and the
MLContext API, so that the user let MLContext know if unregistered
dialects are allowed, so that if needed, you will receive a standard
OperationCompanion transparently.

On UnregisteredOperation's side, now the name is to be passed in a first
parameter list, which actually returns an
`OperationCompanion[UnregisteredOperation]`, implementing the right name
and `apply`.